flexfile class utilities

Functions to help with with the 'flexfile' class type.

is_flexfile() checks if object is of type 'flexfile'

as_flexfile() creates object of type 'flexfile' and checks it against the file specification.

is_flexfile_list() check if the object is a list where all members are of class 'flexfile'.

is_flexfile_family() checks if object is of type 'flexfile_family'

one_of_flexfile_family() checks if object is of a type within the FlexFile family. This means is of type flexfile, quantityreport, maintrepair, or techdatareport.

which_flexfile_family() returns which type within the family the object is. Results are flexfile, quantityreport, maintrepair, or techdatareport.

is_flexfile(x)

as_flexfile(
  x,
  names_case = c("snake_case", "data_model"),
  allocated = FALSE,
  rolledup = FALSE,
  .drop_optional = TRUE,
  .show_check = TRUE
)

is_flexfile_list(x)

is_flexfile_family(x)

one_of_flexfile_family(x)

which_flexfile_family(x)

Arguments

x

An object to test or coerce to type 'flexfile'.

names_case

Case of the object being passed in.

allocated

Logical whether the flexfile has been allocated.

rolledup

Logical whether the flexfile WBS has been rolled up.

.drop_optional

Logical whether to drop optional columns or not.

.show_check

Logical whether or not to show results from the check against the file specification.

Details

The flexfile class has the following attributes.

allocated

Logical. Have the allocations been applied to the ActualCostHourData table.

rolledup

Logical. Has the WBS roll up been applied to the ActualCostHourData and ForecastAtCompletionCostHourData tables.