quantityreport class utilities

Functions to help with with the 'quantityreport' class type.

is_quantityreport() checks if object is of type 'quantityreport'.

as_quantityreport() creates object of type 'quantityreport' and checks it against the file specification.

is_quantityreport_list() check if the object is a list where all members are of class 'quantityreport'.

is_quantityreport(x)

as_quantityreport(
  x,
  names_case = c("snake_case", "data_model"),
  .drop_optional = TRUE,
  .show_check = TRUE
)

is_quantityreport_list(x)

Arguments

x

An object to test or coerce to type 'quantityreport'.

names_case

Case of the object being passed in.

.drop_optional

Logical whether to drop optional columns or not.

.show_check

Logical whether or not to show results from the check against the file specification.