Stack list of FlexFiles or Quantity Reports into one list

This function is depreciated because stacking individual JSON tables is no longer recommended. See vignette("importing-flexfile") for the preferred workflow. If the functionality is still required, use unnest_df() instead.

stack_ff() reads in a list of lists of FlexFile or Quantity Report tibbles and returns a single list of stacked tibbles. The data list should not mix report type (i.e., it should contain all FlexFiles or all Quantity Reports).

This is a thin wrapper around unnest_df().

stack_ff(data)

Arguments

data

A list of FlexFile or Quantity Report tibbles converted from JSON format.

Value

A list of stacked tibbles.

Examples

if (FALSE) {
folder <- system.file("extdata", package = "flexample")

flexfiles <- read_folder(folder, read_flexfile) %>%
  listindex_to_col() %>%
  stack_ff()
}